feat: Add request counter middleware
The `middleware/index.js` file has been modified to include a new middleware function `calculateRequest` that increments the `numberOfRequest` variable for each request. This middleware is used in the `/calculate` route to return the total number of requests made. This addition improves the functionality of the Express server by providing a way to track the number of requests.