refactor: Add fetch API to get animal data
The `fetchApi/index.html` file has been added to the repository. It includes a script that uses the Fetch API to make a request to the "https://fakerapi.it/api/v1/persons" endpoint and logs the response data to the console. This addition improves the functionality of the project by demonstrating how to use the Fetch API to retrieve data from an API.